# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CreatorOS API
  version: 0.10.0
  description: REST contract for the CreatorOS modular monolith.
servers:
  - url: http://localhost:8080
    description: Local API
security: []
...
paths:
...
  /api/v1/campaigns:
    post:
      operationId: createCampaign
      summary: Create a new multi-step campaign draft
      tags: [Campaigns]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignRequest"
      responses:
        "201":
          description: Campaign created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
        "422":
          $ref: "#/components/responses/ValidationFailed"
    get:
      operationId: listCampaigns
      summary: List campaigns owned by the authenticated client
      tags: [Campaigns]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: List of campaigns.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignListResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
...
  /api/v1/admin/overview:
    get:
      operationId: getAdminOverview
      summary: Retrieve aggregate platform metrics and status counts
      tags: [Administration]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Platform overview metrics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverviewResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/admin/users:
    get:
      operationId: listAdminUsers
      summary: List platform users with role and status filtering
      tags: [Administration]
      security:
        - sessionCookie: []
      parameters:
        - name: role
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated user list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/admin/users/{userID}:
    get:
      operationId: getAdminUser
      summary: Get details for a specific user
      tags: [Administration]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/AdminUserID"
      responses:
        "200":
          description: User detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/admin/users/{userID}/status:
    patch:
      operationId: updateAdminUserStatus
      summary: Update user account status (active, disabled)
      tags: [Administration]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/AdminUserID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminUserStatusRequest"
      responses:
        "200":
          description: User status updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/admin/users/{userID}/roles:
    put:
      operationId: updateAdminUserRoles
      summary: Update user assigned roles
      tags: [Administration]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/AdminUserID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminUserRolesRequest"
      responses:
        "200":
          description: User roles updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/admin/orders:
    get:
      operationId: listAdminOrders
      summary: Oversight list of all platform orders with dispute indicators
      tags: [Administration]
      security:
        - sessionCookie: []
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated admin orders.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderListResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/admin/campaigns:
    get:
      operationId: listAdminCampaigns
      summary: Oversight list of all platform campaigns
      tags: [Administration]
      security:
        - sessionCookie: []
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated admin campaigns.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCampaignListResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/admin/disputes:
    get:
      operationId: listAdminDisputes
      summary: List order disputes
      tags: [Administration]
      security:
        - sessionCookie: []
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated order disputes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDisputeListResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/admin/disputes/{orderID}/resolve:
    post:
      operationId: resolveAdminDispute
      summary: Adjudicate an order dispute (refund client, creator payout, or dismiss)
      tags: [Administration]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/AdminOrderID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveAdminDisputeRequest"
      responses:
        "200":
          description: Dispute resolved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/admin/finance/overview:
    get:
      operationId: getAdminFinanceOverview
      summary: Platform financial metrics including escrow, GMV, commissions, and payout queues
      tags: [Administration]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Finance overview metrics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFinanceResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/admin/categories:
    get:
      operationId: listAdminCategories
      summary: List all service categories including inactive ones
      tags: [Administration]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Category list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCategoryListResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
    post:
      operationId: createAdminCategory
      summary: Create a new marketplace category
      tags: [Administration]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminCategoryRequest"
      responses:
        "201":
          description: Category created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/admin/categories/{categoryID}:
    put:
      operationId: updateAdminCategory
      summary: Update category translations, sort order, or active flag
      tags: [Administration]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/AdminCategoryID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminCategoryRequest"
      responses:
        "200":
          description: Category updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/admin/audit-logs:
    get:
      operationId: listAdminAuditLogs
      summary: List append-only audit trail logs
      tags: [Administration]
      security:
        - sessionCookie: []
      parameters:
        - name: action
          in: query
          schema: { type: string }
        - name: resource_type
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Paginated audit logs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuditLogListResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
  /api/v1/admin/announcements:
    get:
      operationId: listAdminAnnouncements
      summary: List platform broadcast announcements
      tags: [Administration]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Announcements list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAnnouncementListResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
    post:
      operationId: createAdminAnnouncement
      summary: Create a new platform announcement
      tags: [Administration]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminAnnouncementRequest"
      responses:
        "201":
          description: Announcement created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAnnouncementResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenAction"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: creatoros_session
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Must match the creatoros_csrf cookie bound to the current session.
      schema:
        type: string
...
    AdminUserID:
      name: userID
      in: path
      required: true
      schema: { type: string, format: uuid }
    AdminCategoryID:
      name: categoryID
      in: path
      required: true
      schema: { type: string, format: uuid }
    AdminOrderID:
      name: orderID
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: The JSON body is malformed, too large, contains unknown fields, or is not a single object.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationFailed:
      description: One or more request fields are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Unauthenticated:
      description: A valid active session is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ForbiddenAction:
      description: The authenticated role or permission cannot perform this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    CampaignVisibility:
      type: string
      enum:
        - public
        - private
    CampaignRequirementInput:
      type: object
      additionalProperties: false
      properties:
        platform_id:
          type: [string, "null"]
          format: uuid
        category_id:
          type: [string, "null"]
          format: uuid
        min_followers:
          type: integer
          format: int64
          minimum: 0
        min_engagement_rate:
          type: number
          minimum: 0
          maximum: 100
        deliverable_type:
          type: string
          maxLength: 80
        quantity:
          type: integer
          minimum: 1
          maximum: 100
    CreateCampaignRequest:
      type: object
      additionalProperties: false
      required:
        - title
        - description
        - budget_minor
        - currency
        - target_creators
        - deadline
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 160
        description:
          type: string
          minLength: 10
          maxLength: 4000
        budget_minor:
          type: integer
          format: int64
          minimum: 1
        currency:
          type: string
          enum: [IDR, MYR, USD]
        target_creators:
          type: integer
          minimum: 1
          maximum: 100
        deadline:
          type: string
          format: date-time
        visibility:
          $ref: "#/components/schemas/CampaignVisibility"
        requirements:
          type: array
          items:
            $ref: "#/components/schemas/CampaignRequirementInput"
...
    CampaignResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Campaign"
    CampaignListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Campaign"
...
    AdminSuccessResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [success]
          properties:
            success:
              type: boolean
              example: true
...
    AdminOverviewResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AdminOverview"
...
    AdminUserResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AdminUser"
    AdminUserListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total: { type: integer }
    UpdateAdminUserStatusRequest:
      type: object
      required: [status, reason]
      properties:
        status: { type: string, enum: [active, disabled] }
        reason: { type: string, minLength: 1 }
    UpdateAdminUserRolesRequest:
      type: object
      required: [roles]
      properties:
        roles:
          type: array
          items: { type: string }
          minItems: 1
...
    AdminOrderListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminOrder"
        total: { type: integer }
...
    AdminCampaignListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminCampaign"
        total: { type: integer }
...
    AdminDisputeListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminDispute"
        total: { type: integer }
    ResolveAdminDisputeRequest:
      type: object
      required: [resolution, notes]
      properties:
        resolution:
          type: string
          enum: [client_refund, creator_payout, dismiss]
        notes: { type: string, minLength: 1 }
...
    AdminFinanceResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AdminFinanceSummary"
...
    AdminCategoryResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AdminCategory"
    AdminCategoryListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminCategory"
    CreateAdminCategoryRequest:
      type: object
      required: [slug, name_id, name_en, name_ms]
      properties:
        slug: { type: string, minLength: 1 }
        name_id: { type: string, minLength: 1 }
        name_en: { type: string, minLength: 1 }
        name_ms: { type: string, minLength: 1 }
        sort_order: { type: integer }
        is_active: { type: boolean }
    UpdateAdminCategoryRequest:
      type: object
      required: [name_id, name_en, name_ms]
      properties:
        name_id: { type: string, minLength: 1 }
        name_en: { type: string, minLength: 1 }
        name_ms: { type: string, minLength: 1 }
        sort_order: { type: integer }
        is_active: { type: boolean }
...
    AdminAuditLogListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminAuditLog"
        total: { type: integer }
...
    CreateAdminAnnouncementRequest:
      type: object
      required: [title, body]
      properties:
        title: { type: string, minLength: 1 }
        body: { type: string, minLength: 1 }
        target_role: { type: string }
    AdminAnnouncementResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AdminAnnouncement"
    AdminAnnouncementListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminAnnouncement"